Moved Code Into Class + Simplified - Jack Vega
Moved all app code into a new GCSApp class and based everything there. I also made it a little easier for other to read and understand what is going on for the tkinter code (although longer).

I also noted that the power_sub, power_motors, and power_servos functions should all be moved to the SM (State Machine) and the GCS should import them from there so communication between the GCS and so there can be less communication between the GCS and other modules.