Sonar Object Sensing Update
Updated sonar code to allow for sensing of objects that don't return a strong intensity. Objects that return an intensity below the threshold will now be sensed by comparing the change of intensity from the intensity at a previous gradian. Logic to deal with sensing if the object continued across multiple gradians was also added. Fixed some minor logical errors in the code as well.